Fixed an issue where cronjob is using today's data
Should be using yesterday's data as that's the time when we have a full
day of data where as we do not have a full day of data today.

Fixed last comment that was forgotten about..

Turns out some files were not added

I was editing the generated copy as oppose to the working copy.