Topcrashers makeover
Add tablesorter and styling to topcrashers to bring it in line with other tables
across the webapp. simplify the styling a little bit for trends to change the
color of the whole cell and vertically center the trending arrow. Also expand
the table to fill in the whole panel rather than condensing everything to the
left side and move the version(s) out of the report title and into the panel
title elements.